Extract converter to a separate file
This will allow it to be used as a base class for real converters.
