+++ sr/comp/scorer/__init__.py
@@ -1,5 +1,7 @@
 from .app import app
+from .converter import Converter
 
 __all__ = (
     'app',
+    'Converter',
 )
